CORE-5001 Use transaction consistently in singular in packages
