+++ libs/ledger/ledger-common-impl/src/main/java/net/corda/ledger/common/impl/transaction/package-info.java
@@ -1,5 +1,5 @@
 @Export
-package net.corda.ledger.common.impl.transactions;
+package net.corda.ledger.common.impl.transaction;
 
 import org.osgi.annotation.bundle.Export;
