+++ libs/ledger/ledger-common-impl/src/main/kotlin/net/corda/ledger/common/impl/transaction/CpkIdentifier.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.common.impl.transactions
+package net.corda.ledger.common.impl.transaction
 
 import net.corda.v5.base.annotations.CordaSerializable
