+++ libs/ledger/ledger-common-impl/src/main/kotlin/net/corda/ledger/common/impl/transaction/PrivacySaltImpl.kt
@@ -1,9 +1,9 @@
-package net.corda.ledger.common.impl.transactions
+package net.corda.ledger.common.impl.transaction
 
 import net.corda.v5.base.annotations.CordaSerializable
 import net.corda.v5.base.types.OpaqueBytes
 import net.corda.v5.crypto.SecureHash
-import net.corda.v5.ledger.common.transactions.PrivacySalt
+import net.corda.v5.ledger.common.transaction.PrivacySalt
 
 /**
  * A privacy salt is required to compute nonces per transaction component in order to ensure that an adversary cannot
