+++ libs/ledger/ledger-consensual-impl/src/main/kotlin/net/corda/ledger/consensual/impl/transaction/ConsensualTransactionMetaData.kt
@@ -1,3 +1,3 @@
-package net.corda.ledger.consensual.impl.transactions
+package net.corda.ledger.consensual.impl.transaction
 
 const val TRANSACTION_META_DATA_CONSENSUAL_LEDGER_VERSION = "0.001"
